refactor: improve code to follow Rust idioms #2

+++ src/card_gen/human.rs
@@ -12,7 +12,7 @@ use super::builder::CardBuilder;
     // - DrawCardFromDeck
     // --------------------------------------------------------
     #[allow(non_snake_case)]
-    pub fn HM_001(card_json: &CardJson, count: usize) -> Card {
+    pub fn HM_001(card_json: &CardJson, count: i32) -> Card {
         CardBuilder::new(card_json)
         .unwrap()
         .add_effect(DrawEffect { count: 2 })
@@ -28,31 +28,31 @@ use super::builder::CardBuilder;
     }
 
     #[allow(non_snake_case)]
-    pub fn HM_002(card_json: &CardJson, count: usize) -> Card {
+    pub fn HM_002(card_json: &CardJson, count: i32) -> Card {
         HM_001(card_json, count)
     }
 
     #[allow(non_snake_case)]
-    pub fn HM_003(card_json: &CardJson, count: usize) -> Card {
+    pub fn HM_003(card_json: &CardJson, count: i32) -> Card {
         HM_001(card_json, count)
     }
     #[allow(non_snake_case)]
-    pub fn HM_004(card_json: &CardJson, count: usize) -> Card {
+    pub fn HM_004(card_json: &CardJson, count: i32) -> Card {
         HM_001(card_json, count)
     }
     #[allow(non_snake_case)]
-    pub fn HM_005(card_json: &CardJson, count: usize) -> Card {
+    pub fn HM_005(card_json: &CardJson, count: i32) -> Card {
         HM_001(card_json, count)
     }
     #[allow(non_snake_case)]
-    pub fn HM_006(card_json: &CardJson, count: usize) -> Card {
+    pub fn HM_006(card_json: &CardJson, count: i32) -> Card {
         HM_001(card_json, count)
     }
     #[allow(non_snake_case)]
-    pub fn HM_007(card_json: &CardJson, count: usize) -> Card {
+    pub fn HM_007(card_json: &CardJson, count: i32) -> Card {
         HM_001(card_json, count)
     }
     #[allow(non_snake_case)]
-    pub fn HM_008(card_json: &CardJson, count: usize) -> Card {
+    pub fn HM_008(card_json: &CardJson, count: i32) -> Card {
         HM_001(card_json, count)
     }
